scripts/run_dev.py
#!/usr/bin/env python
"""
개발 서버 실행 스크립트
"""
import os
import sys

# 프로젝트 루트를 Python 경로에 추가
project_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
sys.path.insert(0, project_root)

if __name__ == "__main__":
    import uvicorn
    from dotenv import load_dotenv
    
    # 환경 변수 로드
    load_dotenv()
    
    uvicorn.run(
        "app.main:app",
        host="0.0.0.0",
        port=8000,
        reload=True,
        log_level="debug",
    )

