pyproject.toml
[project]
name = "ai-vibe-coding-test-worker"
version = "0.1.0"
description = "AI Vibe Coding Test Worker - LangGraph 기반 AI 평가 시스템"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # 기존 의존성
    "google>=3.0.0",
    "ipykernel>=6.31.0",
    "langchain-anthropic>=0.3.22",
    "langchain-google-genai>=3.2.0",
    "langchain-google-vertexai>=2.1.2",
    "langchain-openai>=0.3.35",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.11",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "typing-extensions>=4.15.0",
    "uvicorn[standard]>=0.38.0",
    # FastAPI & Web
    "fastapi>=0.109.0",
    "python-multipart>=0.0.6",
    # Pydantic Settings
    "pydantic-settings>=2.1.0",
    # Database - PostgreSQL
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    # Redis
    "redis>=5.0.0",
    # HTTP Client
    "httpx>=0.26.0",
    # Utils
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
]
